EvmCollectorEtl-master/requirements.txt
web3
eth-abi
eth-utils
python-dateutil
click
ethereum-dasm
base58
requests
pytest
timeout-decorator
google-cloud-pubsub
sqlalchemy
pg8000
pymongo
pyetherchain
QueryStateLib
prometheus-client
python-dotenv
